BC-451 move news repo to shared
