+++ apps/server/src/modules/news/controller/news.controller.spec.ts
@@ -1,8 +1,8 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { LoggerModule } from '@src/core/logger/logger.module';
 import { AuthorizationModule } from '@src/modules/authorization/authorization.module';
+import { NewsRepo } from '@shared/repo';
 import { NewsController } from './news.controller';
-import { NewsRepo } from '../repo/news.repo';
 import { NewsUc } from '../uc';
 
 describe('NewsController', () => {
