+++ apps/server/src/modules/news/news.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { LoggerModule } from '@src/core/logger/logger.module';
 import { AuthorizationModule } from '@src/modules/authorization/authorization.module';
+import { NewsRepo } from '@shared/repo';
 import { NewsUc } from './uc/news.uc';
-import { NewsRepo } from './repo/news.repo';
 import { NewsController } from './controller/news.controller';
 import { TeamNewsController } from './controller/team-news.controller';
