+++ apps/server/src/modules/news/uc/news.uc.spec.ts
@@ -6,7 +6,7 @@ import { UnauthorizedException } from '@nestjs/common';
 import { NewsTargetModel, ICreateNews } from '@shared/domain';
 
 import { AuthorizationService } from '@src/modules/authorization/authorization.service';
-import { NewsRepo } from '../repo/news.repo';
+import { NewsRepo } from '@shared/repo/news';
 import { NewsUc } from './news.uc';
 
 describe('NewsUc', () => {
