+++ apps/server/src/shared/repo/index.ts
@@ -4,8 +4,6 @@
  * It is forbidden to reuse existing ones from different layers.
  */
 
-export * from './query'; // TODO remove
-export { Scope } from './scope'; // TODO remove
-
-export * from './lesson';
 export * from './files';
+export * from './lesson';
+export * from './news';
