+++ apps/server/src/shared/repo/news/index.ts
@@ -1,0 +1,2 @@
+export * from './news.repo';
+export * from './news-target-filter';
